refactor(types): move 6 wire types (Page, BriefingResponse, ContradictionItem, NarrativeResponse, MemoryDetail, PendingRevision) to origin-types
- Add origin-types/src/pages.rs, briefing.rs, narrative.rs (new modules)
- Extend origin-types/src/responses.rs with MemoryDetail + PendingRevision
- Declare + re-export new modules in origin-types/src/lib.rs
- Replace struct definitions in origin-core with pub use re-exports
- Convert Page::new_id() to free function new_page_id() in origin-core::pages
  (can't add impl on foreign type; update all call sites in distill, emergence,
  eval/shared, memory_routes)

All four crates compile clean. No existing call sites broken.
